Clean repeated code and strings, added PushReportCommon
